Add useApplicationData for refactoring
