+++ src/components/Application.js
@@ -6,155 +6,21 @@ import "components/Application.scss";
 import DayList from "./DayList";
 import Appointment from "./Appointment";
 import { getAppointmentsForDay, getInterview, getInterviewersForDay } from "helpers/selectors";
-
-// const days = [
-//   {
-//     id: 1,
-//     name: "Monday",
-//     spots: 2,
-//   },
-//   {
-//     id: 2,
-//     name: "Tuesday",
-//     spots: 5,
-//   },
-//   {
-//     id: 3,
-//     name: "Wednesday",
-//     spots: 0,
-//   },
-// ];
-
-// const appointments = [
-//   {
-//     id: 1,
-//     time: "12pm",
-//   },
-//   {
-//     id: 2,
-//     time: "1pm",
-//     interview: {
-//       student: "Lydia Miller-Jones",
-//       interviewer:{
-//         id: 3,
-//         name: "Sylvia Palmer",
-//         avatar: "https://i.imgur.com/LpaY82x.png",
-//       }
-//     }
-//   },
-//   {
-//     id: 3,
-//     time: "2pm",
-//   },
-//   {
-//     id: 4,
-//     time: "3pm",
-//     interview: {
-//       student: "Archie Andrews",
-//       interviewer:{
-//         id: 4,
-//         name: "Cohana Roy",
-//         avatar: "https://i.imgur.com/FK8V841.jpg",
-//       }
-//     }
-//   },
-//   {
-//     id: 5,
-//     time: "4pm",
-//   }
-// ];
+import useApplicationData from "hooks/useApplicationData";
 
 
 
 export default function Application(props) {
-  // const [day, setDay] = useState('Monday');
-  // const [days, setDays] = useState(['test']);
-  // const [appointmets, setAppointments] = useState({});
+ 
+  const {
+    state,
+    setDay,
+    bookInterview,
+    cancelInterview
+  } = useApplicationData();
 
-
-  const [state, setState] = useState({
-    day: "Monday",
-    days: [],
-    appointments: {},
-    interviewers: {}
-
-  });
-
-  let dailyAppointments = [];
-  let dayInterviewers = [];
-
-  const setDay = day => setState({ ...state, day });
-  // const setDays = days => setState(prev => ({ ...prev, days }));
-
-
-  useEffect(() => {
-    Promise.all([
-      axios.get('/api/days'),
-      axios.get('/api/appointments'),
-      axios.get('/api/interviewers')
-    ]).then(all => {
-      // console.log('Fetched data: ', "days", all[0].data, "app", all[1].data, "int", all[2].data);
-      setState(prev => ({ ...prev, days: all[0].data, appointments: all[1].data, interviewers: all[2].data }));
-    })
-      .catch(err => console.log(err.message)
-      );
-  }, []);
-
-  dailyAppointments = getAppointmentsForDay(state, state.day);
-  dayInterviewers = getInterviewersForDay(state, state.day);
-
-  function bookInterview(id, interview) {
-
-    const appointment = {
-      ...state.appointments[id],
-      interview: { ...interview }
-    };
-    const appointments = {
-      ...state.appointments,
-      [id]: appointment
-    };
-
-    return (
-
-      axios.put(`api/appointments/${id}`, appointment)
-        .then((res) => {
-          console.log(res);
-          setState({
-            ...state,
-            appointments
-          });
-
-        })
-        // .catch(err => console.log(err.message))
-    );
-
-  }
-
-  function cancelInterview(id) {
-    const appointment = {
-      ...state.appointments[id],
-      interview: null
-    };
-    const appointments = {
-      ...state.appointments,
-      [id]: appointment
-    };
-
-    return (
-
-      axios.delete(`api/appointments/${id}`, appointment)
-        .then((res) => {
-          console.log(res);
-          setState({
-            ...state,
-            appointments
-          });
-
-        })
-        // .catch(err => console.log(err.message))
-    );
-
-  }
+  const dailyAppointments = getAppointmentsForDay(state, state.day);
+  const dayInterviewers = getInterviewersForDay(state, state.day);
 
   return (
     <main className="layout">
